requirements.txt
gitpython
matplotlib
aiofiles
tqdm
argparse